+++ .changeset/real-sloths-accept.md
@@ -1,0 +1,5 @@
+---
+'@open-editor/client': patch
+---
+
+add framework react15
