+++ packages/client/src/resolve/framework/svelte.ts
@@ -1,0 +1,14 @@
+import { getComponentNameByFile } from '../util';
+
+export function resolveSourceFromSvelte(
+  meta?: { loc: { file?: string } } | null,
+) {
+  if (!meta) return [];
+
+  return [
+    {
+      name: getComponentNameByFile(meta.loc.file, 'svelte'),
+      file: meta.loc.file,
+    },
+  ];
+}
