+++ packages/client/src/resolve/index.ts
@@ -1,0 +1,1 @@
+export * from './resolveSource';
