+++ playground/rollup-react/package.json
@@ -6,8 +6,8 @@
     "dev": "rollup -c -w --environment NODE_ENV:development"
   },
   "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react": "^15.7.0",
+    "react-dom": "^15.7.0"
   },
   "devDependencies": {
     "@babel/core": "^7.22.10",
@@ -19,12 +19,12 @@
     "@rollup/plugin-commonjs": "^25.0.4",
     "@rollup/plugin-node-resolve": "^15.1.0",
     "@rollup/plugin-replace": "^5.0.2",
-    "@types/react": "^18.2.21",
-    "@types/react-dom": "^18.2.7",
+    "@types/react": "^15.5.0",
+    "@types/react-dom": "^15.5.0",
     "rollup": "^3.28.1",
     "rollup-plugin-live-server": "^2.0.0",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-svg": "^2.0.0",
-    "typescript": "^5.1.6"
+    "typescript": "^2.2.0"
   }
 }
