+++ playground/rollup-react/src/tree/SubTree1.tsx
@@ -1,0 +1,4 @@
+import SubTree2 from './SubTree2';
+export default function SubTree1(): React.JSX.Element {
+  return <SubTree2 />;
+}
