+++ playground/rollup-react/src/tree/SubTree2.tsx
@@ -1,0 +1,4 @@
+import SubTree3 from './SubTree3';
+export default function SubTree2(): React.JSX.Element {
+  return <SubTree3 />;
+}
