+++ playground/rollup-react/src/tree/SubTree3.tsx
@@ -1,0 +1,3 @@
+export default function SubTree3(): React.JSX.Element {
+  return <button>test tree</button>;
+}
