+++ playground/rollup-react/src/tree/index.tsx
@@ -1,0 +1,4 @@
+import SubTree1 from './SubTree1';
+export default function Tree(): React.JSX.Element {
+  return <SubTree1 />;
+}
