+++ playground/vite-qwik/src/Notes.tsx
@@ -1,4 +1,5 @@
 import { component$ } from '@builder.io/qwik';
+import Tree from './tree';
 
 const Notes = component$(() => {
   return (
@@ -7,6 +8,10 @@ const Notes = component$(() => {
         press hotkey ⌨️ <kbd>option ⌥</kbd> + <kbd>command ⌘</kbd> +{' '}
         <kbd>O</kbd>, then click the HTML element you wish to inspect.
       </p>
+      <p>
+        press hotkey ⌨️ <kbd>command ⌘</kbd> + 🖱 click, show component tree.
+      </p>
+      <Tree />
       <p>
         press hotkey ⌨️ <kbd>esc</kbd> or 🖱 right-click to exit inspect.
       </p>
