+++ playground/webpack-next/src/tree/SubTree2.tsx
@@ -1,0 +1,4 @@
+import SubTree3 from './SubTree3';
+export default function SubTree2() {
+  return <SubTree3 />;
+}
