+++ playground/webpack-vue/src/components/tree/SubTree1.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import SubTree2 from './SubTree2.vue';
+</script>
+
+<template>
+  <SubTree2 />
+</template>
