+++ playground/webpack-vue/src/components/tree/SubTree2.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import SubTree3 from './SubTree3.vue';
+</script>
+
+<template>
+  <SubTree3 />
+</template>
