+++ playground/webpack-vue/src/components/tree/index.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import SubTree1 from './SubTree1.vue';
+</script>
+
+<template>
+  <SubTree1 />
+</template>
